Rework autogpt
Closes #10
