+++ actions/index.js
@@ -1,0 +1,3 @@
+export {interactWithPage} from './interactWithPage.js';
+export {goToLink} from './goToLink.js';
+export {findInPage} from './findInPage.js';
